+++ crypto/protocol/src/protocol_transport/subscribe_message.rs
@@ -1,7 +1,6 @@
 use serde::{Deserialize, Serialize};
-use serde_big_array::BigArray;
+use sp_core::{sr25519, Pair};
 use subxt::utils::AccountId32;
-use subxt_signer::sr25519;
 
 /// A message sent by a party when initiating a websocket connection to participate
 /// in the signing or DKG protcol
@@ -11,31 +10,22 @@ pub struct SubscribeMessage {
     /// Protocol session identifier
     pub session_id: String,
     /// Public key of connecting party
-    pub public_key: [u8; 32],
+    pub public_key: sr25519::Public,
     /// Signature to authenticate connecting party
-    #[serde(with = "BigArray")]
-    pub signature: [u8; 64],
+    pub signature: sr25519::Signature,
 }
 
 impl SubscribeMessage {
-    pub fn new(session_id: &str, sk: &sr25519::Keypair) -> Self {
-        let signature = sk.sign(session_id.as_bytes());
-        Self {
-            session_id: session_id.to_owned(),
-            public_key: sk.public_key().0,
-            signature: signature.0,
-        }
+    pub fn new(session_id: &str, pair: &sr25519::Pair) -> Self {
+        let signature = pair.sign(session_id.as_bytes());
+        Self { session_id: session_id.to_owned(), public_key: pair.public(), signature }
     }
 
     pub fn account_id(&self) -> AccountId32 {
-        self.public_key.into()
+        self.public_key.0.into()
     }
 
     pub fn verify(&self) -> bool {
-        sr25519::verify(
-            &sr25519::Signature(self.signature),
-            self.session_id.as_bytes(),
-            &sr25519::PublicKey(self.public_key),
-        )
+        sr25519::Pair::verify(&self.signature, self.session_id.as_bytes(), &self.public_key)
     }
 }